Move JDBC Meta data API tests into TestKit.
We also simplify the configuration for these tests with extra
capabilities and clean up PlainSQLTest.
