+++ slick-testkit/src/main/resources/testkit-reference.conf
@@ -27,6 +27,7 @@ testkit {
     ${testPackage}.JoinTest
     ${testPackage}.MainTest
     ${testPackage}.JdbcMapperTest
+    ${testPackage}.JdbcMetaTest
     ${testPackage}.MetaModelTest
     ${testPackage}.RelationalMapperTest
     ${testPackage}.RelationalMiscTest
